Stub out bpf_tap_packet_xxx functions
The bpf_tap_packet_in and bpf_tap_packet_out functions appear to
be almost the same as bpf_tap_in and bpf_tap_out. The difference is
they take in a kern_packet_t which I believe is a Skywalk related
data structure. The two missing functions are simply stubbed along
with all the other Skywalk related functions which are not part of the
open source Apple released.